pyproject.toml
[project]
name = "skillforge"
version = "0.1.0"
description = "MCP server for managing and iteratively optimizing Skills"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillforge"]

[tool.hatch.build.targets.wheel.force-include]
"src/skillforge/guide" = "skillforge/guide"

[project.scripts]
skillforge = "skillforge.server:main"
